feat(components): organize/declutter shared components folder
