--- packages/blockchain-wallet-v4-frontend/src/components/Lockbox/index.js
@@ -1,3 +1,0 @@
-import AppManager from './AppManager'
-
-export { AppManager }
