+++ packages/blockchain-wallet-v4-frontend/src/components/QRCode/Reader/index.tsx
