+++ packages/blockchain-wallet-v4-frontend/src/components/QRCode/Wrapper/index.js
