+++ packages/blockchain-wallet-v4-frontend/src/modals/Borrow/BorrowForm/Summary/index.tsx
@@ -1,15 +1,17 @@
+import { FormattedMessage } from 'react-intl'
+import React from 'react'
+import styled from 'styled-components'
+
 import { BorrowFormValuesType } from 'data/types'
 import {
   coinToString,
   fiatToString,
   formatFiat
 } from 'blockchain-wallet-v4/src/exchange/currency'
-import { FormattedMessage } from 'react-intl'
 import { OfferType, RatesType } from 'core/types'
-import { TableRow, Title, Value } from 'components/Borrow'
 import { Text, TooltipHost, TooltipIcon } from 'blockchain-info-components'
-import React from 'react'
-import styled from 'styled-components'
+
+import { TableRow, Title, Value } from '../../components'
 
 type Props = {
   collateral: number
