+++ packages/blockchain-wallet-v4-frontend/src/modals/Btc/RequestBtc/template.success.js
@@ -21,7 +21,7 @@ import { invalidAmountMax, invalidAmountMin } from './validation'
 import { model } from 'data'
 import { required } from 'services/FormHelper'
 import CopyClipboard from 'components/CopyClipboard'
-import QRCodeWrapper from 'components/QRCodeWrapper'
+import QRCodeWrapper from 'components/QRCode/Wrapper'
 
 const AddressContainer = styled.div`
   display: flex;
