+++ packages/blockchain-wallet-v4-frontend/src/modals/Eth/SendEth/FirstStep/template.success.js
@@ -54,7 +54,7 @@ import MinFeeForRetryInvalid from './MinFeeForRetryInvalid'
 import MnemonicRequiredForCustodySend from 'components/Send/RecoveryPhrase'
 import PriorityFeeLink from './PriorityFeeLink'
 import PropTypes from 'prop-types'
-import QRCodeCapture from 'components/QRCodeCapture'
+import QRCodeCapture from 'components/QRCode/Capture'
 import React from 'react'
 import RegularFeeLink from './RegularFeeLink'
 import styled from 'styled-components'
