+++ packages/blockchain-wallet-v4-frontend/src/modals/Lockbox/components/index.js
