+++ packages/blockchain-wallet-v4-frontend/src/modals/QRCode/template.js
@@ -1,6 +1,6 @@
 import { FormattedMessage } from 'react-intl'
 import PropTypes from 'prop-types'
-import QRCodeWrapper from 'components/QRCodeWrapper'
+import QRCodeWrapper from 'components/QRCode/Wrapper'
 import React from 'react'
 import styled from 'styled-components'
