+++ packages/blockchain-wallet-v4-frontend/src/modals/Settings/TwoStepGoogleAuthenticator/template.js
@@ -1,6 +1,6 @@
 import { Field, reduxForm } from 'redux-form'
 import { FormattedMessage } from 'react-intl'
-import QRCodeWrapper from 'components/QRCodeWrapper'
+import QRCodeWrapper from 'components/QRCode/Wrapper'
 import React from 'react'
 import styled from 'styled-components'
