+++ packages/blockchain-wallet-v4-frontend/src/scenes/Airdrops/components/EmailRequired.js
