+++ packages/blockchain-wallet-v4-frontend/src/scenes/Settings/Addresses/QrCodeAddressInfo/index.tsx
@@ -2,7 +2,7 @@ import { FormattedMessage } from 'react-intl'
 import { Text } from 'blockchain-info-components'
 import CoinDisplay from 'components/Display/CoinDisplay'
 import media from 'services/ResponsiveService'
-import QRCodeWrapper from 'components/QRCodeWrapper'
+import QRCodeWrapper from 'components/QRCode/Wrapper'
 import React from 'react'
 import styled from 'styled-components'
